requirements.txt
transformers==4.36.2
torch 
torchvision 
torchaudio
python-dotenv
datasets[audio]==2.16.0
soundfile==0.12.1
sentencepiece==0.1.99
accelerate==0.25.0